# internal/adapters/http/docs/swagger.yaml
basePath: /api
definitions:
  audit.AuditActionDTO:
    properties:
      action:
        description: '"user.create"'
        type: string
      category:
        description: '"user"'
        type: string
      description:
        description: '"Create new user"'
        type: string
      label:
        description: '"创建用户"'
        type: string
      operation:
        description: '"create"'
        type: string
    type: object
  audit.AuditActionsResponseDTO:
    properties:
      actions:
        description: 操作定义列表
        items:
          $ref: '#/definitions/audit.AuditActionDTO'
        type: array
      categories:
        description: 分类选项
        items:
          $ref: '#/definitions/audit.CategoryOptionDTO'
        type: array
      operations:
        description: 操作类型选项
        items:
          $ref: '#/definitions/audit.OperationTypeDTO'
        type: array
    type: object
  audit.AuditDTO:
    properties:
      action:
        type: string
      created_at:
        type: string
      details:
        type: string
      id:
        type: integer
      ip_address:
        type: string
      resource:
        type: string
      status:
        type: string
      user_agent:
        type: string
      user_id:
        type: integer
    type: object
  audit.CategoryOptionDTO:
    properties:
      label:
        description: '"用户管理"'
        type: string
      value:
        description: '"user"'
        type: string
    type: object
  audit.OperationTypeDTO:
    properties:
      label:
        description: '"创建"'
        type: string
      value:
        description: '"create"'
        type: string
    type: object
...
  cache.CacheInfoDTO:
    properties:
      db_size:
        description: DBSize 当前 key 数量（仅计算应用前缀下的）
        type: integer
      key_prefix:
        description: KeyPrefix 应用使用的 key 前缀
        type: string
      memory_usage:
        description: MemoryUsage 内存使用量（如果可用）
        type: string
      redis_version:
        description: RedisVersion Redis 版本
        type: string
    type: object
  cache.CacheKeyDTO:
    properties:
      key:
        description: Key 完整 key 名称
        type: string
      ttl:
        description: TTL 剩余过期时间（秒），-1 表示永不过期，-2 表示不存在
        type: integer
      type:
        description: Type key 类型（string, ReJSON-RL, hash 等）
        type: string
    type: object
  cache.CacheValueDTO:
    properties:
      key:
        description: Key 完整 key 名称
        type: string
      ttl:
        description: TTL 剩余过期时间（秒），-1 表示永不过期，-2 表示不存在
        type: integer
      type:
        description: Type key 类型（string, ReJSON-RL, hash 等）
        type: string
      value:
        description: Value key 的值（JSON 格式）
        items:
          type: integer
        type: array
    type: object
  cache.DeleteResultDTO:
    properties:
      deleted_count:
        description: DeletedCount 删除的 key 数量
        type: integer
    type: object
  cache.ScanKeysResultDTO:
    properties:
      cursor:
        description: Cursor 下一个游标（用于分页），为 "0" 表示结束
        type: string
      keys:
        description: Keys key 列表
        items:
          $ref: '#/definitions/cache.CacheKeyDTO'
        type: array
      total_scanned:
        description: TotalScanned 本次扫描的 key 数量
        type: integer
    type: object
...
  response.DataResponse-array_routes_OperationDefinition:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        description: 响应数据
        items:
          $ref: '#/definitions/routes.OperationDefinition'
        type: array
      error:
        description: 错误详情（仅失败时）
      message:
        description: 消息描述
        type: string
    type: object
...
  response.DataResponse-audit_AuditActionsResponseDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/audit.AuditActionsResponseDTO'
        description: 响应数据
      error:
        description: 错误详情（仅失败时）
      message:
        description: 消息描述
        type: string
    type: object
  response.DataResponse-audit_AuditDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/audit.AuditDTO'
        description: 响应数据
      error:
        description: 错误详情（仅失败时）
      message:
        description: 消息描述
        type: string
    type: object
...
  response.DataResponse-cache_CacheInfoDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/cache.CacheInfoDTO'
        description: 响应数据
      error:
        description: 错误详情（仅失败时）
      message:
        description: 消息描述
        type: string
    type: object
  response.DataResponse-cache_CacheValueDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/cache.CacheValueDTO'
        description: 响应数据
      error:
        description: 错误详情（仅失败时）
      message:
        description: 消息描述
        type: string
    type: object
  response.DataResponse-cache_DeleteResultDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/cache.DeleteResultDTO'
        description: 响应数据
      error:
        description: 错误详情（仅失败时）
      message:
        description: 消息描述
        type: string
    type: object
  response.DataResponse-cache_ScanKeysResultDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/cache.ScanKeysResultDTO'
        description: 响应数据
      error:
        description: 错误详情（仅失败时）
      message:
        description: 消息描述
        type: string
    type: object
...
  response.ErrorDetail:
    properties:
      code:
        description: 业务错误码（小写下划线）
        type: string
      details:
        description: 额外详情（如验证错误列表）
      message:
        description: 错误消息
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/response.ErrorDetail'
    type: object
...
  response.PagedResponse-audit_AuditDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        description: 数据列表
        items:
          $ref: '#/definitions/audit.AuditDTO'
        type: array
      message:
        description: 消息描述
        type: string
      meta:
        allOf:
        - $ref: '#/definitions/response.PaginationMeta'
        description: 分页信息
    type: object
...
  response.PaginationMeta:
    properties:
      has_more:
        description: 是否有下一页
        type: boolean
      page:
        description: 当前页码
        type: integer
      per_page:
        description: 每页数量
        type: integer
      total:
        description: 总记录数
        type: integer
      total_pages:
        description: 总页数
        type: integer
      warning:
        description: 页码越界警告
        type: string
    type: object
...
  routes.OperationDefinition:
    properties:
      code:
        description: 操作代码，如 sys:users:create
        type: string
      description:
        description: 描述
        type: string
      group:
        description: Swagger 分组
        type: string
      identifier:
        description: 标识符，如 create
        type: string
      label:
        description: 中文标签
        type: string
      scope:
        description: Scope，如 sys
        type: string
      type:
        description: 类型，如 users
        type: string
    type: object
...
host: localhost:40012
info:
  contact:
    email: lwmacct@icloud.com
    name: API Support
    url: https://github.com/lwmacct/251117-go-ddd-template/issues
  description: 基于 DDD+CQRS 架构的 Go Web 应用 API 文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/lwmacct/251117-go-ddd-template
  title: Go DDD Template API
  version: 1.0.3
paths:
  /api/admin/audit:
    get:
      consumes:
      - application/json
      description: 分页获取审计日志，支持按用户、操作、资源、状态、时间范围筛选
      parameters:
      - description: Action 操作类型过滤（语义化标识，如 setting.update）
        in: query
        name: action
        type: string
      - description: EndDate 结束时间（RFC3339 格式）
        in: query
        name: end_date
        type: string
      - default: 20
        description: Limit 每页数量，默认 20，最大 1000
        in: query
        maximum: 1000
        minimum: 1
        name: limit
        type: integer
      - default: 1
        description: Page 页码，从 1 开始
        in: query
        minimum: 1
        name: page
        type: integer
      - description: Resource 资源分类过滤（如 setting, user）
        in: query
        name: resource
        type: string
      - description: StartDate 开始时间（RFC3339 格式）
        in: query
        name: start_date
        type: string
      - description: Status 状态过滤
        enum:
        - success
        - failure
        in: query
        name: status
        type: string
      - description: UserID 按用户 ID 过滤
        in: query
        name: user_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 审计日志列表
          schema:
            $ref: '#/definitions/response.PagedResponse-audit_AuditDTO'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 审计日志列表
      tags:
      - Admin - Audit Log
  /api/admin/audit/{id}:
    get:
      consumes:
      - application/json
      description: 根据日志ID获取审计日志详细信息
      parameters:
      - description: 日志ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 日志详情
          schema:
            $ref: '#/definitions/response.DataResponse-audit_AuditDTO'
        "400":
          description: 无效的日志ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 日志不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 审计日志详情
      tags:
      - Admin - Audit Log
  /api/admin/audit/actions:
    get:
      consumes:
      - application/json
      description: 获取所有审计操作的定义、分类和操作类型，供前端筛选器使用
      produces:
      - application/json
      responses:
        "200":
          description: 审计操作定义
          schema:
            $ref: '#/definitions/response.DataResponse-audit_AuditActionsResponseDTO'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 审计操作定义
      tags:
      - Admin - Audit Log
  /api/admin/cache/info:
    get:
      consumes:
      - application/json
      description: 查看缓存状态信息（类似 redis-cli INFO）
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.DataResponse-cache_CacheInfoDTO'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 缓存信息
      tags:
      - System
  /api/admin/cache/key:
    delete:
      consumes:
      - application/json
      description: 删除指定的单个 key（类似 redis-cli DEL）
      parameters:
      - description: Key 完整的 key 名称（含前缀）
        in: query
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.DataResponse-cache_DeleteResultDTO'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 删除缓存键
      tags:
      - System
    get:
      consumes:
      - application/json
      description: 获取指定 key 的完整信息和值（类似 redis-cli GET/JSON.GET）
      parameters:
      - description: Key 完整的 key 名称（含前缀）
        in: query
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.DataResponse-cache_CacheValueDTO'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Key 不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取缓存值
      tags:
      - System
  /api/admin/cache/keys:
    delete:
      consumes:
      - application/json
      description: 批量删除匹配 pattern 的所有 keys
      parameters:
      - description: Pattern 匹配模式（不含应用前缀，如 "setting:*"）
        in: query
        name: pattern
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.DataResponse-cache_DeleteResultDTO'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 批量删除缓存
      tags:
      - System
    get:
      consumes:
      - application/json
      description: 按 pattern 扫描缓存 keys（类似 redis-cli SCAN）
      parameters:
      - description: Cursor 游标（用于分页），首次查询传 "0" 或留空
        in: query
        name: cursor
        type: string
      - description: Limit 每次返回的最大数量
        in: query
        maximum: 1000
        minimum: 1
        name: limit
        type: integer
      - description: Pattern 匹配模式（不含应用前缀，如 "setting:*"）
        in: query
        name: pattern
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.DataResponse-cache_ScanKeysResultDTO'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 扫描缓存键
      tags:
      - System
  /api/admin/operations:
    get:
      consumes:
      - application/json
      description: 返回所有可用的操作定义，供前端权限配置使用
      produces:
      - application/json
      responses:
        "200":
          description: 操作列表
          schema:
            $ref: '#/definitions/response.DataResponse-array_routes_OperationDefinition'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 操作列表
      tags:
      - System
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer token authentication. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"